+++ app/Http/Requests/Analytics/StoreDashboardWidgetRequest.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Http\Requests\Analytics;
+
+class StoreDashboardWidgetRequest extends DashboardWidgetRequest {}