+++ app/Http/Requests/Analytics/StoreKpiMetricRequest.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Http\Requests\Analytics;
+
+class StoreKpiMetricRequest extends KpiMetricRequest {}