+++ app/Http/Requests/Analytics/UpdateKpiTargetRequest.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Http\Requests\Analytics;
+
+class UpdateKpiTargetRequest extends KpiTargetRequest {}