+++ app/Http/Requests/Analytics/UpdateUserDashboardPreferenceRequest.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Http\Requests\Analytics;
+
+class UpdateUserDashboardPreferenceRequest extends DashboardPreferenceRequest {}